renaming to pybitup
