--- pybit/__init__.py
@@ -1,6 +1,0 @@
-import pybit.metropolis_hastings_algorithms
-import pybit.post_process
-import pybit.inference_problem
-import pybit.distributions
-import pybit.sample_dist
-import pybit.bayesian_inference
