+++ testCases/heat_conduction/run.py
@@ -1,5 +1,5 @@
 import heat_conduction 
-import pybit
+import pybitup
 
 case_name = "heat_conduction"
 input_file_name = "{}.json".format(case_name) 
@@ -7,6 +7,6 @@ input_file_name = "{}.json".format(case_name)
 
 heat_conduction_model = heat_conduction.HeatConduction()
 
-post_dist = pybit.sample_dist.SolveProblem(input_file_name)
+post_dist = pybitup.sample_dist.SolveProblem(input_file_name)
 post_dist.sample(heat_conduction_model)
 post_dist.post_process_dist()
