+++ testCases/pyrolysis/one_reaction_pyrolysis.py
@@ -2,7 +2,7 @@ from scipy import special
 import numpy as np
 import math	
 
-from pybit import bayesian_inference as bi
+from pybitup import bayesian_inference as bi
 
 class OneReactionPyrolysis(bi.Model):
