+++ testCases/pyrolysis/run_rescaled.py
@@ -1,6 +1,6 @@
 import one_reaction_pyrolysis 
 import numpy as np
-import pybit
+import pybitup
 import matplotlib.pyplot as plt
 
 case_name = "one_reaction_pyrolysis_rescaled"
@@ -35,6 +35,6 @@ print(X)
 
 
 
-post_dist = pybit.sample_dist.SolveProblem(input_file_name)
+post_dist = pybitup.sample_dist.SolveProblem(input_file_name)
 post_dist.sample(pyro_model)
 post_dist.post_process_dist()
