+++ testCases/pyrolysis_parallel_fran_model/ito_sde/2_param/fake_data/run_ito_sde_2param.py
@@ -4,7 +4,7 @@ from pyrolysis_general.src.pyrolysis import PyrolysisParallel
 import parallel_reaction_2param
 
 # Packages for stochastic inference
-import pybit
+import pybitup
 
 case_name = "pyrolysis_parallel_2param"
 
@@ -13,6 +13,6 @@ input_file_name = "{}.json".format(case_name)
 pyrolysis_model = parallel_reaction_2param.SetParallelReaction()
 
 
-post_dist = pybit.sample_dist.SolveProblem(input_file_name)
+post_dist = pybitup.sample_dist.SolveProblem(input_file_name)
 #post_dist.sample(pyrolysis_model)
 post_dist.post_process_dist()
