+++ testCases/pyrolysis_parallel_fran_model/run_2react.py
@@ -4,7 +4,7 @@ from pyrolysis_general.src.pyrolysis import PyrolysisParallel
 import parallel_reaction
 
 # Packages for stochastic inference
-import pybit
+import pybitup
 import matplotlib.pyplot as plt
 
 # Python packages
@@ -15,7 +15,7 @@ input_file_name = "{}.json".format(case_name)
 
 pyrolysis_model = parallel_reaction.SetParallelReaction()
 
-my_dist = pybit.sample_dist.SolveProblem(input_file_name)
+my_dist = pybitup.sample_dist.SolveProblem(input_file_name)
 my_dist.sample(pyrolysis_model)
 my_dist.post_process_dist()
