+++ testCases/spring_model/run.py
@@ -1,5 +1,5 @@
 import spring_model 
-import pybit
+import pybitup
 
 #case_name = "spring_model"
 case_name = "spring_model"
@@ -7,6 +7,6 @@ input_file_name = "{}.json".format(case_name)
 
 spring_model = spring_model.SpringModel(name=case_name)
 
-post_dist = pybit.sample_dist.SolveProblem(input_file_name)
+post_dist = pybitup.sample_dist.SolveProblem(input_file_name)
 post_dist.sample(spring_model)
 post_dist.post_process_dist()
